requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
dotenv==0.9.9
idna==3.10
lxml==6.0.0
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
typing_extensions==4.14.1
urllib3==2.5.0
